normalize image ID and name to fix stale re-scan
